backend/requirements.txt
fastapi==0.61.1
motor==2.2.0
pydantic[email]
python-decouple==3.3
uvicorn==0.11.8